src/bot_dispatcher/requirements.txt
langchain
pydantic
boto
botocore
aws-requests-auth
